+++ src/TestSuite/TestCase.php
@@ -13,9 +13,12 @@
  */
 namespace MeTools\TestSuite;
 
+use Cake\Filesystem\Folder;
 use Cake\TestSuite\TestCase as CakeTestCase;
+use Exception;
 use MeTools\TestSuite\MockTrait;
-use MeTools\TestSuite\Traits\TestCaseTrait;
+use Tools\ReflectionTrait;
+use Tools\TestSuite\TestCaseTrait;
 
 /**
  * TestCase class
@@ -23,6 +26,7 @@ use MeTools\TestSuite\Traits\TestCaseTrait;
 abstract class TestCase extends CakeTestCase
 {
     use MockTrait;
+    use ReflectionTrait;
     use TestCaseTrait;
 
     /**
@@ -45,5 +49,54 @@ abstract class TestCase extends CakeTestCase
         parent::tearDown();
 
         safe_unlink_recursive(LOGS);
+    }
+
+    /**
+     * Internal method to get a log full path
+     * @param string $filename Log filename
+     * @return string
+     * @since 2.16.10
+     */
+    protected function getLogFullPath($filename)
+    {
+        if (!pathinfo($filename, PATHINFO_EXTENSION)) {
+            $filename .= '.log';
+        }
+
+        return Folder::isAbsolute($filename) ? $filename : LOGS . $filename;
+    }
+
+    /**
+     * Asserts log file contents
+     * @param string $expectedContent The expected contents
+     * @param string $filename Log filename
+     * @param string $message The failure message that will be appended to the
+     *  generated message
+     * @return void
+     * @uses getLogFullPath()
+     */
+    public function assertLogContains($expectedContent, $filename, $message = '')
+    {
+        $filename = $this->getLogFullPath($filename);
+
+        try {
+            is_readable_or_fail($filename);
+            $content = file_get_contents($filename);
+        } catch (Exception $e) {
+            $this->fail($e->getMessage());
+        }
+
+        $this->assertContains($expectedContent, $content, $message);
+    }
+
+    /**
+     * Deletes a log file
+     * @param string $filename Log filename
+     * @return void
+     * @uses getLogFullPath()
+     */
+    public function deleteLog($filename)
+    {
+        safe_unlink($this->getLogFullPath($filename));
     }
 }