+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * `InstallShell` has been replaced with console commands. Every method of the
     previous class is now a `MeTools\Command\Install` class;
 * `MockTrait` moved from `MeTools\TestSuite\Traits` to `MeTools\TestSuite`;
+* removed `TestCaseTrait`, because now all the necessary methods are in the
+    `TestCase` class;
 * `cakephp-assets`, `jquery`, `bootstrap`, `bootstrap-datetimepicker`,
     `fancy-box` and `fortawesome--fontawesome-free` are no longer required
     packages for Composer (`cakephp-assets` is required for developing), but